requirements.txt
Pillow>=10.0
cairosvg>=2.7
requests>=2.31
